Refactor error hints into focused modules
